requirements.txt
discord.py
python-dotenv
pillow
aiosqlite
